# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Furniture API Documentation",
        "description": "API documentation for the Furniture E-commerce Application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Current API Server"
        }
    ],
    "paths": {
...
        "/api/bank/login": {
            "post": {
                "operationId": "57c97b7e0072c2611df86bfa9d1320dc",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "account_number": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bank login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/bank/account": {
            "get": {
                "operationId": "377814d0078427a0248eedaa3a506e15",
                "responses": {
                    "200": {
                        "description": "Bank account details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
...
        "/api/favorites": {
            "get": {
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "responses": {
                    "200": {
                        "description": "List favorites"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "operationId": "d426a8b38354581cb4082ac4bc62e22f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Added to favorites"
                    },
                    "400": {
                        "description": "Already in favorites"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites/{product_id}": {
            "delete": {
                "operationId": "58857f1be6ec489f33a8a46c43b2cbdc",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed from favorites"
                    },
                    "404": {
                        "description": "Favorite not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites/toggle": {
            "post": {
                "operationId": "c4cb21639a868dcedea199141c109a18",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Favorite status toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/favorites/check/{product_id}": {
            "get": {
                "operationId": "3c0639aa617e7f7625fc6c045e8df01b",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}